Test all mixers in test_mixer
